Fix incorrect sign extension on PCIe bus number
git-svn-id: http://casvn/svn/octeon/bdk/branches/thunder@107980 c83883e8-9f28-0410-baf1-f2025811e16a

+++ libbdk-hal/bdk-pcie.c
@@ -822,7 +822,7 @@ static uint64_t __bdk_pcie_build_config_addr(bdk_node_t node, int pcie_port, int
                address. Use the SLI regions instead of ECAMs for config space
                access */
             uint64_t address = bdk_pcie_get_base_address(node, pcie_port, BDK_PCIE_MEM_CONFIG);
-            address += bus << 24;   /* Bus is bits 31:24 */
+            address += (uint64_t)bus << 24;   /* Bus is bits 31:24 */
             address += dev << 19;   /* device+func is bits 23:16 */
             address += fn << 16;
             address += reg;         /* Offset is bits 11:0 */
